move LTRFeatureLoggerTransformerFactory to solr.response.transform
(this is where the other transformers live and so when in the future solr/contrib/ltr graduates to solr/core then users therefore need not update their configs)
