+++ solr/contrib/ltr/src/test-files/solr/collection1/conf/solrconfig-ltr.xml
@@ -36,7 +36,7 @@
   enclosed between brackets (in this case [fv]). In order to get the feature
   vector you will have to specify that you want the field (e.g., fl="*,[fv]) -->
  <transformer name="fv"
-  class="org.apache.solr.ltr.ranking.LTRFeatureLoggerTransformerFactory" />
+  class="org.apache.solr.response.transform.LTRFeatureLoggerTransformerFactory" />
 
  <updateHandler class="solr.DirectUpdateHandler2">
   <autoCommit>
